refactor(traces): route the run drawer's thread through the Trace Explorer
ScenarioRunDetailDrawer opened the legacy trace view by rendering TraceDetails
inside a second Drawer.Root of its own, held open by local state — the pattern
dev/docs/best_practices/drawers.md exists to rule out, and its own comment said
"managed via local state". "Open thread" now opens the Trace Explorer drawer
with mode=conversation, which is the same view through the drawer system
everything else in the product uses. The run drawer's other trace action,
"View conversation in Trace Explorer", is unchanged.

That was the last thing holding the legacy trace view up, and removing it takes
21 files with it — the whole of components/messages, and TraceDetails plus the
seven files only it reached (Evaluations, Events, PinButton, SequenceDiagram,
SpanTree, Summary, groupEvaluations).

components/traces keeps the eight files that other owners import: the media
parts and AddParticipants that traces-v2 itself renders, RenderInputOutput
(six consumers), SpanDetails, EvaluationStatusItem, ThreadMapping and
TracesMapping. A reachability walk from every importer outside the two
directories is what drew the line; the eight are shared code, not leftovers.

TraceWithGuardrail moves to server/traces/types.ts. It is the shape the trace
list read returns — a Trace plus the guardrail that blocked it and whether it
carries annotations — and it lived in a React component, so two server files
imported a component for a type. The frontend-boundary guard had a case
tolerating that import because types are erased; with its subject gone the
assertion would have passed vacuously, so it moves to a real remaining
type-only component import and now pins both halves: the import has to still
be there AND still be excluded.
