--- platform/app/src/components/messages/Conversation.tsx
@@ -1,181 +1,0 @@
-import {
-  Alert,
-  Box,
-  Grid,
-  GridItem,
-  Skeleton,
-  Text,
-  VStack,
-} from "@chakra-ui/react";
-import { useEffect, useRef } from "react";
-import { useRouter } from "~/utils/compat/next-router";
-import { useOrganizationTeamProject } from "../../hooks/useOrganizationTeamProject";
-import { useTraceDetailsState } from "../../hooks/useTraceDetailsState";
-import { api } from "../../utils/api";
-import { Message, TraceMessages } from "./TraceMessages";
-
-export function Conversation({
-  threadId,
-  traceId,
-}: {
-  threadId?: string;
-  traceId: string;
-}) {
-  const router = useRouter();
-  const traceIdParam = (router.query.trace as string) || traceId;
-  const { trace } = useTraceDetailsState(traceIdParam);
-
-  const { project, isPublicRoute } = useOrganizationTeamProject();
-
-  const currentTraceRef = useRef<HTMLDivElement>(null);
-  const threadTraces = api.traces.getTracesByThreadId.useQuery(
-    {
-      projectId: project?.id ?? "",
-      threadId: threadId ?? "",
-    },
-    {
-      enabled: !!project && !!threadId,
-    },
-  );
-
-  const modalTraceId =
-    (router.query.trace as string) ||
-    (threadTraces.data?.length && threadTraces.data?.length > 1
-      ? router.query["drawer.traceId"]
-      : undefined);
-
-  useEffect(() => {
-    if (threadTraces.data && threadTraces.data?.length > 0 && modalTraceId) {
-      // Wait for DOM to render, then scroll to highlighted trace
-      const timeoutId = setTimeout(() => {
-        if (currentTraceRef.current) {
-          currentTraceRef.current.scrollIntoView({
-            behavior: "smooth",
-            block: "center",
-          });
-        }
-      }, 300);
-
-      return () => clearTimeout(timeoutId);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [threadTraces.data, modalTraceId]);
-
-  return (
-    <Box width="full" minWidth="800px" paddingX={6}>
-      {threadTraces.data && threadTraces.data.length > 50 && (
-        <Alert.Root status="info">
-          <Alert.Indicator />
-          <Alert.Content>
-            <Alert.Title>
-              Over 50 messages found in thread, showing only the last 50.
-            </Alert.Title>
-          </Alert.Content>
-        </Alert.Root>
-      )}
-      <VStack align="start" width="full" gap={0}>
-        {!!threadId || trace.data ? (
-          <>
-            {threadId ? (
-              threadTraces.data ? (
-                threadTraces.data
-                  .slice(Math.max(0, threadTraces.data.length - 50))
-                  .map((trace, index) => (
-                    <TraceMessages
-                      key={trace.trace_id}
-                      trace={trace}
-                      index={
-                        threadTraces.data.length === 1
-                          ? "only"
-                          : index === 0
-                            ? "first"
-                            : index === threadTraces.data.length - 1
-                              ? "last"
-                              : "other"
-                      }
-                      ref={
-                        trace.trace_id == (modalTraceId || traceId)
-                          ? currentTraceRef
-                          : undefined
-                      }
-                      highlighted={trace.trace_id == modalTraceId}
-                    />
-                  ))
-              ) : threadTraces.error && !isPublicRoute ? (
-                <Box maxWidth="800px" paddingTop={8} paddingBottom={4}>
-                  <Text color="red.500">
-                    Something went wrong trying to load previous messages
-                  </Text>
-                </Box>
-              ) : null
-            ) : null}
-            {trace.data &&
-              (!threadTraces.data || threadTraces.data.length === 0) && (
-                <TraceMessages
-                  trace={trace.data}
-                  highlighted={!!modalTraceId}
-                  index="only"
-                  loadingMore={threadTraces.isFetching}
-                />
-              )}
-          </>
-        ) : trace.isLoading ? (
-          <Box maxWidth="1000px" width="full">
-            <Grid templateColumns="repeat(4, 1fr)">
-              <GridItem colSpan={4}>
-                <Box
-                  position="relative"
-                  border="1px solid"
-                  borderColor="border"
-                  marginRight={5}
-                  paddingLeft={10}
-                  paddingRight={10}
-                  paddingY={4}
-                  background="bg.panel"
-                  borderRadius="lg"
-                  minHeight="524px"
-                >
-                  <Message
-                    author=""
-                    avatar={
-                      <Skeleton
-                        borderRadius="full"
-                        minWidth="36px"
-                        minHeight="36px"
-                      />
-                    }
-                    paddingTop="20px"
-                  >
-                    <Box paddingY="6px" marginBottom="62px" maxWidth="90%">
-                      <VStack gap={4} width="full" align="start">
-                        <Skeleton width="720px" maxWidth="100%" height="20px" />
-                        <Skeleton width="720px" maxWidth="100%" height="20px" />
-                      </VStack>
-                    </Box>
-                  </Message>
-                  <Message
-                    author=""
-                    avatar={
-                      <Skeleton
-                        borderRadius="full"
-                        minWidth="36px"
-                        minHeight="36px"
-                      />
-                    }
-                  >
-                    <Box paddingY="6px" marginBottom="62px" maxWidth="90%">
-                      <VStack gap={4} width="full" align="start">
-                        <Skeleton width="720px" maxWidth="100%" height="20px" />
-                        <Skeleton width="720px" maxWidth="100%" height="20px" />
-                      </VStack>
-                    </Box>
-                  </Message>
-                </Box>
-              </GridItem>
-            </Grid>
-          </Box>
-        ) : null}
-      </VStack>
-    </Box>
-  );
-}
