--- platform/app/src/components/traces/Evaluations.tsx
@@ -1,237 +1,0 @@
-import { Box, HStack, Text, VStack } from "@chakra-ui/react";
-import type { ReactNode } from "react";
-import { useState } from "react";
-import type { Project } from "~/generated/prisma/client";
-import type { ElasticSearchEvaluation } from "~/server/tracer/types";
-import { evaluationPassed } from "../checks/EvaluationStatus";
-import { Link } from "../ui/link";
-import { EvaluationStatusItem } from "./EvaluationStatusItem";
-import {
-  type EvaluationGroup,
-  groupEvaluationsByEvaluator,
-} from "./groupEvaluations";
-
-interface TraceEval {
-  project?: Project;
-  traceId: string;
-  evaluations?: ElasticSearchEvaluation[];
-}
-
-/** Indentation matching the width of the evaluation status icon. */
-const EVALUATION_STATUS_INDENT = "22px";
-
-function EvaluationGroupEntry({ group }: { group: EvaluationGroup }) {
-  const [isExpanded, setIsExpanded] = useState(false);
-
-  return (
-    <Box width="full">
-      <EvaluationStatusItem check={group.latest} />
-      {group.hasPreviousRuns && (
-        <HStack paddingLeft={EVALUATION_STATUS_INDENT} marginTop={1}>
-          <Text
-            fontSize="xs"
-            color="fg.subtle"
-            cursor="pointer"
-            _hover={{ textDecoration: "underline" }}
-            onClick={() => setIsExpanded(!isExpanded)}
-            role="button"
-            aria-expanded={isExpanded}
-          >
-            +{group.previousRunCount} previous
-          </Text>
-        </HStack>
-      )}
-      {isExpanded &&
-        group.runs.slice(1).map((run) => (
-          <Box
-            key={run.evaluation_id}
-            paddingLeft={EVALUATION_STATUS_INDENT}
-            marginTop={2}
-            borderLeftWidth="2px"
-            borderLeftColor="border.subtle"
-            marginLeft="4px"
-          >
-            <EvaluationStatusItem check={run} />
-          </Box>
-        ))}
-    </Box>
-  );
-}
-
-/** Renders a list of evaluation groups with dividers between entries. */
-function EvaluationGroupList({
-  groups,
-  emptyState,
-}: {
-  groups: EvaluationGroup[];
-  emptyState: ReactNode;
-}) {
-  if (groups.length === 0) return <>{emptyState}</>;
-
-  return (
-    <VStack align="start" gap={0} width="full">
-      {groups.map((group, index) => (
-        <Box key={group.groupKey} width="full">
-          {index > 0 && (
-            <Box
-              borderTopWidth="1px"
-              borderTopColor="border.subtle"
-              marginY={3}
-            />
-          )}
-          <EvaluationGroupEntry group={group} />
-        </Box>
-      ))}
-    </VStack>
-  );
-}
-
-export function Evaluations(trace: TraceEval & { anyGuardrails: boolean }) {
-  const evaluations = trace.evaluations?.filter((x) => !x.is_guardrail);
-  const groups = groupEvaluationsByEvaluator(evaluations);
-
-  return (
-    <EvaluationGroupList
-      groups={groups}
-      emptyState={
-        <Text>
-          No evaluations ran for this message.{" "}
-          {trace.anyGuardrails ? (
-            "Evaluations are skipped if guardrails completely blocked the message."
-          ) : (
-            <>
-              Setup evaluations{" "}
-              <Link
-                href={`/${trace.project?.slug}/online-evaluations`}
-                textDecoration="underline"
-              >
-                here
-              </Link>
-              .
-            </>
-          )}
-        </Text>
-      }
-    />
-  );
-}
-
-export const Guardrails = (trace: TraceEval) => {
-  const guardrails = trace.evaluations?.filter((x) => x.is_guardrail);
-  const groups = groupEvaluationsByEvaluator(guardrails);
-
-  return (
-    <EvaluationGroupList
-      groups={groups}
-      emptyState={
-        <Text>
-          No guardrails ran for this message. Setup guardrails{" "}
-          <Link
-            href={`/${trace.project?.slug}/online-evaluations`}
-            textDecoration="underline"
-          >
-            here
-          </Link>
-          .
-        </Text>
-      }
-    />
-  );
-};
-
-export const EvaluationsCount = (
-  trace: TraceEval & { countGuardrails?: boolean },
-) => {
-  const evaluations = trace.countGuardrails
-    ? trace.evaluations?.filter((x) => x.is_guardrail)
-    : trace.evaluations?.filter((x) => !x.is_guardrail);
-
-  const groups = groupEvaluationsByEvaluator(evaluations);
-
-  // A fail is a verdict; a crashed evaluator is not (#6835). Keep the two
-  // apart so an infrastructure error never reads as "the output failed".
-  const totalFailed = groups.filter(
-    (group) => evaluationPassed(group.latest) === false,
-  ).length;
-  const totalErrored = groups.filter(
-    (group) => group.latest.status === "error",
-  ).length;
-
-  if (totalFailed > 0 || totalErrored > 0) {
-    if (trace.countGuardrails) {
-      return null;
-    }
-
-    return (
-      <HStack gap={1}>
-        {totalFailed > 0 && (
-          <Text
-            borderRadius={"md"}
-            paddingX={2}
-            backgroundColor={"red.solid"}
-            color={"white"}
-            fontSize={"sm"}
-          >
-            {totalFailed} failed
-          </Text>
-        )}
-        {totalErrored > 0 && (
-          <Text
-            borderRadius={"md"}
-            paddingX={2}
-            backgroundColor={"orange.solid"}
-            color={"white"}
-            fontSize={"sm"}
-          >
-            {totalErrored} errored
-          </Text>
-        )}
-      </HStack>
-    );
-  }
-
-  const totalProcessed = groups.filter(
-    (group) => group.latest.status === "processed",
-  ).length;
-  const total = groups.length;
-
-  if (total === 0) return null;
-
-  return (
-    <Text
-      borderRadius={"md"}
-      paddingX={2}
-      backgroundColor={totalProcessed > 0 ? "green.solid" : "yellow.solid"}
-      color={"white"}
-      fontSize={"sm"}
-    >
-      {totalProcessed > 0 ? totalProcessed : total}
-    </Text>
-  );
-};
-
-export const Blocked = (trace: TraceEval) => {
-  const guardrails = trace.evaluations?.filter((x) => x.is_guardrail);
-  const groups = groupEvaluationsByEvaluator(guardrails);
-
-  // Status-aware: only a guardrail that ran to completion can have blocked —
-  // an errored guardrail's stray `passed: false` is not a block (#6833).
-  const totalBlocked = groups.filter(
-    (group) =>
-      group.latest.status === "processed" && group.latest.passed === false,
-  ).length;
-
-  if (totalBlocked === 0) return null;
-
-  return (
-    <Text
-      borderRadius={"md"}
-      paddingX={2}
-      backgroundColor={"blue.subtle"}
-      color={"blue.fg"}
-      fontSize={"sm"}
-    >
-      {totalBlocked} blocked
-    </Text>
-  );
-};
