+++ platform/app/src/server/traces/clickhouse-trace.service.ts
@@ -2,7 +2,6 @@ import type { ClickHouseClient } from "@clickhouse/client";
 import { HandledError } from "@langwatch/handled-error";
 import { createLogger } from "@langwatch/observability";
 import { getLangWatchTracer } from "langwatch";
-import type { TraceWithGuardrail } from "~/components/messages/MessageCard";
 import type { PrismaClient } from "~/generated/prisma/client";
 import { LLM_PARAMETER_MAP } from "~/prompts/prompt-playground/llmParameterMap";
 import { AnnotationService } from "~/server/annotations/annotation.service";
@@ -60,6 +59,7 @@ import type {
   TopicCountsResult,
   TraceDateField,
   TracesForProjectResult,
+  TraceWithGuardrail,
 } from "./types";
 
 /**
